requirements.txt
google-cloud-dialogflow==2.6.0
python-dotenv==0.18.0
python-telegram-bot==13.7
vk-api==11.9.4
